#1046: extract TCP segmentation builders to frame/tcp_segmentation.rs
frame/mod.rs went from 1,995 prod LOC to 1,678 by relocating
`segment_forwarded_tcp_frames_from_frame` (307 lines, L181-487) and
the thin `segment_forwarded_tcp_frames` XdpDesc-adapter wrapper
(L740-757) into a dedicated child module.

Visibility widened from `pub(super)` (which made them visible to
afxdp via `use self::frame::*;`) to `pub(in crate::afxdp)` directly
on the moved fns, then re-exported through frame/mod.rs at the same
level so afxdp.rs's existing `use self::frame::*;` continues to
surface them at every call site (notably tx/dispatch.rs).

Pure relocation — no production logic touched, no test bodies
modified. The bodies are byte-for-byte identical; only the enclosing
module changed.

LOC distribution after:
- frame/mod.rs:               1,678 (was 1,995) — comfortably below threshold
- frame/tcp_segmentation.rs:    337 (new)

This implements one of the four submodules called out in #1046's
"Clean Sweep" plan (tcp_segmentation/nat64/gre/icmp builders).
NAT64 was already a thin 40-line dispatcher to `crate::nat64` — not
worth a separate file. GRE and ICMP builders are similarly thin and
deferred unless frame/mod.rs grows back toward threshold.

cargo test --release: 865 passed, 0 failed.
